Fixing incompatibility with node-rdkafka new version
In version 1.0.0, we upgrade some core components. But some few points break. This PR fixes all the broken scenarios and remove methods that doesn't exists anymore.

Unfortunately the consumer stream needs to be dropped - for now - the node-rdkafka implementation changes a lot.

I can provide a new way to do this in next versions