Adding fetchRefreshRate option to SchemaRegistry
The strategy that kafka-avro uses to deal with schemas are a full load in memory. This PR adds the possibility to set a refresh rate to schemas fetch - before, the schemas are only loaded when init.

This is not the best strategy at all - cached requests dealing with schemas on demand is a better option. But, to add this option we need to do a major refactoring in the way this lib works and some users are having problems with it.

So, let's add this option for now - the full code refactoring and the new strategy will come in the next major release.

In addition, a better try/catch handle to avoid crashes.

Fixes https://github.com/waldophotos/kafka-avro/issues/41