Fixing avsc failling to parse schemas with the same name (#2)
Deep coping avsc parse options. Avsc will add registry to those options and if we try to parse schema with the same name it will fail if e.g. with fetch all versions enabled, trying to avsc.parse an evolved version of the same schema will throw an exception that there is a duplicate schema in the registry.